+++ crates/runtime/src/prelude/encoding-streams.js
@@ -114,6 +114,11 @@
     }
   }
 
-  globalThis.TextEncoderStream = TextEncoderStream;
-  globalThis.TextDecoderStream = TextDecoderStream;
+  for (const Interface of [TextEncoderStream, TextDecoderStream]) {
+    Object.defineProperty(Interface.prototype, Symbol.toStringTag, {
+      value: Interface.name,
+      configurable: true,
+    });
+    globalThis[Interface.name] = Interface;
+  }
 })();